requirements.txt
# Professional Background Remover Requirements
# Core AI and Image Processing
rembg>=2.0.50
onnxruntime-gpu>=1.15.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Advanced Image Processing
opencv-python>=4.8.0
scikit-image>=0.21.0
scipy>=1.11.0
numpy>=1.24.0
Pillow>=10.0.0
imageio>=2.31.0

# Professional Features
matplotlib>=3.7.0
plotly>=5.15.0
albumentations>=1.3.0

# Performance and GPU Acceleration
numba>=0.57.0

# Color Science and Professional Output
colour-science>=0.4.2

# Batch Processing and Workflow
tqdm>=4.65.0
joblib>=1.3.0

# File Format Support
imageio-ffmpeg>=0.4.8

# Quality Control and Metrics
lpips>=0.1.4
